+++ linetools/analysis/interactive_plot.py
@@ -362,7 +362,7 @@ q        : quit
         self.artists['indices'] = art
         self.artists['initcont'], = a0.plot(wa, self.continuum, color='k', lw=2, ls='dashed', zorder=3)
         self.artists['fl'], = a0.plot(wa, fl, lw=1, color='0.7',
-                                      linestyle='steps-mid')
+                                      drawstyle='steps-mid')
         a0.plot(wa, er, lw=0.5, color='orange')
         m1, = a0.plot([0], [0], 'r', zorder=4, lw=2)
         m2, = a0.plot([0], [0], 'o', mfc='None', mew=2, ms=12, mec='r',